+++ border-atari-env/src/util/test.rs
@@ -45,7 +45,7 @@ impl SubBatch for ObsBatch {
     }
 
     #[inline]
-    fn push(&mut self, i: usize, data: &Self) {
+    fn push(&mut self, i: usize, data: Self) {
         unsafe {
             let src: *const u8 = &data.buf[0];
             let dst: *mut u8 = &mut self.buf[i * self.m];
@@ -100,7 +100,7 @@ impl SubBatch for ActBatch {
     }
 
     #[inline]
-    fn push(&mut self, i: usize, data: &Self) {
+    fn push(&mut self, i: usize, data: Self) {
         unsafe {
             let src: *const u8 = &data.buf[0];
             let dst: *mut u8 = &mut self.buf[i * self.m];