+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 
-## v0.0.6 (20??-??-??)
+## v0.0.7 (20??-??-??)
+
+### Added
+
+Support MLflow tracking (`border-mlflow-tracking`) (https://github.com/taku-y/border/issues/2).
+
+### Changed
+
+* Take `self` in the signature of `push()` method of replay buffer (`border-core`)
+
+## v0.0.6 (2023-09-19)
 
 ### Added